pyproject.toml
[project]
name = "fintra"
version = "0.1.0"
description = "Backend for fintra app"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "prometheus-client>=0.22.1",
    "psycopg>=3.2.9",
    "python-dotenv>=1.1.1",
    "starlette>=0.47.1",
    "uvicorn>=0.35.0",
    "yoyo-migrations>=9.0.0",
]

[dependency-groups]
debug = [
    "debugpy>=1.8.14",
]
test = [
    "httpx>=0.28.1",
    "pytest>=8.4.1",
    "pytest-asyncio>=1.0.0",
    "pytest-env>=1.1.5",
]
